common-utils - fixing testsuite
java.net.URL.getPort() returns -1 if port is not set in the URL,
regardless if the scheme (http / https) is specified or not. I think it
is the best way to fix these tests (non intrusive in the current
codebase and coherent with default Java behaviour, considering the code
in GeonetHttpRequestFactory is supposed to work).