Fix MD language corruption
When tomcat was restarted and the first MD was viewed by a user, the
MD was re-indexed (to update the view counter) in a Quartz thread.
This Quartz thread never had his ServiceContext initialized and that
made the ISO3 to ISO2 conversion unable to convert the codes.

I guess something later on is setting the ServiceContext for the Quartz
thread and the problem doesn't happen anymore, because only the first MD
that is viewed is corrupted.