Fix multi contact deletion
When configuring a field in one of the views of the config-editor.xml
file and this field was present more that once in the MD, the
form-configurator.xml was not inserting the element selected by the field's
xpath, but its direct child. If the element was present only once in the MD,
the behavior was correct and the element selected by the xpath was inserted.

Because the che:CHE_CI_ResponsibleParty was inserted in the editor instead of
the gmd:contact, it was not possible to delete it (its parent is a xlink).