Fix empty SchematronCriteria empty value
When a schematron criteria was of type XPATH or ALWAYS_ACCEPT it didn't
send the right criteria value to the server to be save. Now, if the
criteria type is not using a typeahead control, it uses the criteria type
value to generate the actual value for the criteria replacing
`@@value@@` by the valueui in the criteriaType value field.