sb-451 - forcing xpath selection order
In a xpath.selectNodes() call, you cannot predict the actual order of
the list containing the selected elements. a selectSingleNode() will
return the first object of the previous list, we want to make sure of
the selection order, and we end up doing it by hand, splitting the
xpath.

Tests:
- updated testsuite for AAP (be careful that the behaviour is not
  exactly the same between junit and the live webapp). Even without the
  modification in AapMetadataReport, the newly added test was passing.

- runtime-tested

Note:
the same approach might be necessary for the other xpaths, but let's
see.