avoid last element in page when sorting to be wrong
doc states that compareBottom has to return "N > 0 if the doc's value is sorted before the bottom entry",
so when bottom is null, return is 1.

Also want comparaison for bottom upper case to be.

refactor

do not understand why we should forbid copy (and overwrite) when value is null

doc states that this method is awaiting for an 'howMany' parameter,

and 'Returns the documents in the rage [start .. start+howMany) '