Aggregation / Add terms script support
Elastic provide support of script in aggregation config using painless
language. See https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations-bucket-terms-aggregation.html#search-aggregations-bucket-terms-aggregation-script

For example, the resource identifier is using a predefined scheme eg. `eurostat_v_4258_100_k_gisco-mreg-2021_i_2021_v01_r01` and user wants to build a facet on the first part of the identifier. The script can extract it and build an aggregation on it. Then the configuration needs to indicate which field has to be used using `meta.field` and if a wildcard mode is required.

``json
'facetConfig': {
  "OrgForResource": {
    "terms" : {
      "script": {
        "inline": "doc['resourceIdentifier.code'].size() == 0 ? 'http' : /_/.split(doc['resourceIdentifier.code'].getValue())[0]"
      },
      "exclude": "http.*",
      "size": 10,
      "order" : { "_key" : "asc" }
    },
    "meta": {
      "field": "resourceIdentifier.code",
      "wildcard": true
    }
  },
```