[déjà dans gn] SEO / Set HTML head title and description
Currently HTML head description is set to an empty value. Google is
indexing the cookie warning in GeoNetwork case when no HTML head is found (probably because it is the first HTML text content found).

Use the service metadata configured for the main node (see admin >
settings > record to use for GetCapabilities) and same for portal.

If no service metadata record configured, default title and description
to node name (or site name for the main portal).

Follow up of:
* https://github.com/geonetwork/core-geonetwork/commit/611be2a0f5a514be9fa3b80a14b0dd785c94dba9
* https://github.com/geonetwork/core-geonetwork/pull/4758